Extract circuit breaker action to separate class, small code fixes.
